+++ CHANGELOG.md
@@ -22,6 +22,8 @@
   available inside the container.
 - The `remote status` command will now print the username, realname, and email
   of the logged-in user, if available.
+- OCI mode now supports `--hostname` (requires UTS namespace, therefore this
+  flag will infer `--uts` if running in OCI mode).
 
 ### Bug Fixes